Make crossystem.h more polite and more useful.
This adds a VB_MAX_STRING_PROPERTY for callers that don't
want to guess at how big to make their buffers.

Additionally, it changes the size parameter to VbGetPropertyString()
from int to size_t.

BUG=None
TEST=compile the code
BRANCH=none

Reviewed-on: https://chromium-review.googlesource.com/175039
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>

+++ host/arch/arm/lib/crossystem_arch.c
@@ -5,6 +5,7 @@
 
 #include <stdio.h>
 #include <string.h>
+#include <stddef.h>
 #include <stdlib.h>
 #include <linux/fs.h>
 #include <sys/types.h>
@@ -533,7 +534,8 @@ int VbGetArchPropertyInt(const char* name) {
     return -1;
 }
 
-const char* VbGetArchPropertyString(const char* name, char* dest, int size) {
+const char* VbGetArchPropertyString(const char* name, char* dest,
+                                    size_t size) {
   char *str = NULL;
   char *rv = NULL;
   char *prop = NULL;
